Refactor `nzbToMedia.process` -> `core.processor.nzb.process`
